Split vnf_interface.c into common, LTE and NR implementations
